Move out of copy package to staging
